requirements.txt
alabaster==0.7.12
alembic==1.8.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
awscli==1.27.10
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
Boruta==0.3
botocore==1.29.10
catalogue==1.0.2
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.0
colorama==0.4.4
colorlover==0.3.0
coverage==6.5.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
Cython==0.29.14
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
docker==6.0.1
docopt==0.6.2
docutils==0.16
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flake8==5.0.4
Flask==2.2.2
fonttools==4.38.0
funcy==1.17
future==0.18.2
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.1
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imagesize==1.4.1
imbalanced-learn==0.7.0
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
kmodes==0.12.2
lightgbm==3.3.3
llvmlite==0.37.0
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
missingno==0.5.1
mistune==2.0.4
mlflow==2.0.1
mlxtend==0.19.0
multimethod==1.9
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.54.1
numexpr==2.8.4
numpy==1.20.3
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
pandas-profiling==3.4.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
phik==0.12.2
pickleshare==0.7.5
Pillow==9.3.0
pipreqs==0.4.11
pkgutil_resolve_name==1.3.10
plac==1.1.3
platformdirs==2.5.4
plotly==5.11.0
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.21.9
psutil==5.9.4
pure-eval==0.2.2
pyarrow==10.0.0
pyasn1==0.4.8
pycaret==2.3.10
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.6.0
pyLDAvis==3.3.1
pynndescent==0.5.8
pyod==1.0.6
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyWavelets==1.4.1
pywin32==305
pywinpty==2.0.9
PyYAML==5.4.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.12.1
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
sklearn==0.0.post1
slicer==0.0.7
smart-open==6.2.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
spacy==2.3.8
Sphinx==5.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.44
sqlparse==0.4.3
srsly==1.0.6
stack-data==0.6.1
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.1.0
terminado==0.17.0
textblob==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.12
visions==0.7.5
waitress==2.1.2
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wincertstore==0.2
wordcloud==1.8.2.2
yarg==0.1.9
yellowbrick==1.2.1
zipp==3.10.0
